Added ShortCode functionality which is configured simply by adding a 'ShortCode' section to the MvcConfiguration::append in bootstrap.php
eg:

MvcConfiguration::append(array(
	'AdminPages' => array(

...

	),
	'ShortCodes' => array(
		'testscode' => array(
				'options' => array('controller' => 'speakers', 'action' => 'show'),
				'allow_actions' => array('show'),
				'deny_actions' => '',
		),
	),
));

then you can use the shortcode:

[testscode id=2]

which will be substitued with the results of the venues/show for venue id 2

Allow / Deny
============

if you specify action in the options, then this is the default action,
otherwise the default action is 'shortcode'.

if you specify allow_actions, then the action must match one of the allow_actions
otherwise it will fail with an error

if you specify deny_actions, then the action must not match one of the deny_actions
otherwise it will fail with an error

the allow and deny actions can be a single string or array of strings.
an attempt is made to match each string initially as a direct match, then
as a pcre regular expression.

the default allow_actions deny_actions configuration is that the action
shortcode, or any action begining with 'sc_' are allowed, and no
deny_actions is specified.